Modified DELETE Questionnaires endpoint to use Transactions.
Modified the DELETE Questionnaires endpoint to use Transactions to delete the Questionnaire and the Questions and Answers linked to it. Also, modified the get_db() function in db.py to return the MongoClient object from the connection to the database instead of the database object. All the endpoints then use the Flask config object to retrieve the DATABASE_DB_NAME to get the database object for all the commands sent to the database.