Modify 'auth' and 'user' endpoints.
Modified, primarily, the code for the 'auth' and 'user' endpoints to be able to work with the frontend on registering and loging in users.\n* Added the 'user_exists' function in db.py to check in the database if a user already exists with the given username or email.\n* Added more error messages to user.py that inform the user if the username or the email address are already in use by another user.\n* Fixed a bug in auth.py that would cause an uncaught error to happen when the user trying to log in doesn't exist. The /auth endpoint now successfully handles invalid users returning an error that informs the user that either their username or password are wrong.\n Changed HTTP 403 codes in user.py and auth.py to more suitable 401 codes.\n Added CORS support in endpoint.py so browsers can allow the frontend to communicate with the backend.