Added JWT authentication
Added JWT authentication using Refresh Tokens and Access Tokens. When the user logs in, the server provides the client with a Refresh Token through a cookie. This Refresh Token is saved as an 'active' session in the database. The client also gets an Access Token that is sent through the JSON in the response. This Access Token must be used in the client's next request. Every time a request is sent, the client gets a new Refresh and Access Token. If the Refresh Token is stolen and reused, the server will delete all session from the database that correspond to this user. This will cause the user to log out from all sessions/devices.